fix(client): break circular imports in agent-selection hook extraction
Extracts useAgents/useResolvedAgentId/getAgentName out of
components/AgentSelector.tsx into hooks/useAgents.ts (mirroring
hooks/useEmbeddedAgents.ts), and switches AgentSelector.tsx's
components/agents import to direct file paths instead of the barrel.
Both changes remove circular dependency edges flagged by dependency-cruiser
in CI (bun run lint:deps), which is not covered by bun run test.
