+++ packages/client/src/components/worktrees/CreateWorktreeForm.tsx
@@ -3,7 +3,8 @@ import { useForm } from 'react-hook-form';
 import { valibotResolver } from '@hookform/resolvers/valibot';
 import { useMutation, useQuery, useQueryClient } from '@tanstack/react-query';
 import { FormField, Input, Textarea } from '../ui/FormField';
-import { UnifiedAgentSelector, useResolvedAgentId, useResolvedEmbeddedAgentId } from '../AgentSelector';
+import { UnifiedAgentSelector, useResolvedEmbeddedAgentId } from '../AgentSelector';
+import { useResolvedAgentId } from '../../hooks/useAgents';
 import { Spinner } from '../ui/Spinner';
 import type { CreateWorktreeFormData } from '../../schemas/worktree-form';
 import { CreateWorktreeFormSchema } from '../../schemas/worktree-form';
